test: Move CLI app helper to main helper implementation
This allows for access from all test suites.
